pydantic fix
