+++ api/src/opentrons/protocol_api/core/engine/well.py
@@ -3,12 +3,16 @@ from typing import Optional, Union
 
 from opentrons_shared_data.labware.constants import WELL_NAME_PATTERN
 
+from opentrons.types import Point
+
 from opentrons.protocol_engine import WellLocation, WellOrigin, WellOffset
 from opentrons.protocol_engine import commands as cmd
 from opentrons.protocol_engine.clients import SyncClient as EngineClient
 from opentrons.protocols.api_support.util import UnsupportedAPIError
-
-from opentrons.types import Point, SimulatedProbeResult, LiquidTrackingType
+from opentrons.protocol_engine.types.liquid_level_detection import (
+    SimulatedProbeResult,
+    LiquidTrackingType,
+)
 
 from . import point_calculations
 from . import stringify
