+++ api/src/opentrons/protocol_api/core/instrument.py
@@ -11,6 +11,7 @@ from opentrons.protocols.api_support.util import FlowRates
 from opentrons.protocols.advanced_control.transfers.common import TransferTipPolicyV2
 from opentrons.protocol_api._nozzle_layout import NozzleLayout
 from opentrons.protocol_api._liquid import LiquidClass
+from opentrons.protocol_engine.types.liquid_level_detection import LiquidTrackingType
 
 from ..disposal_locations import TrashBin, WasteChute
 from .well import WellCoreType
@@ -427,7 +428,7 @@ class AbstractInstrument(ABC, Generic[WellCoreType, LabwareCoreType]):
     @abstractmethod
     def liquid_probe_without_recovery(
         self, well_core: WellCoreType, loc: types.Location
-    ) -> types.LiquidTrackingType:
+    ) -> LiquidTrackingType:
         """Do a liquid probe to find the level of the liquid in the well."""
         ...
