+++ api/src/opentrons/protocol_api/core/well.py
@@ -3,7 +3,8 @@
 from abc import ABC, abstractmethod
 from typing import TypeVar, Optional, Union
 
-from opentrons.types import Point, LiquidTrackingType
+from opentrons.types import Point
+from opentrons.protocol_engine.types.liquid_level_detection import LiquidTrackingType
 
 from .._liquid import Liquid
