+++ api/src/opentrons/protocol_engine/state/frustum_helpers.py
@@ -5,7 +5,10 @@ from math import isclose
 
 from ..errors.exceptions import InvalidLiquidHeightFound
 
-from opentrons.types import LiquidTrackingType, SimulatedProbeResult
+from opentrons.protocol_engine.types.liquid_level_detection import (
+    LiquidTrackingType,
+    SimulatedProbeResult,
+)
 from opentrons_shared_data.labware.labware_definition import (
     InnerWellGeometry,
     WellSegment,
@@ -361,7 +364,6 @@ def find_volume_at_well_height(
 ) -> LiquidTrackingType:
     """Find the volume within a well, at a known height."""
     # comparisons with SimulatedProbeResult objects aren't meaningful, just return
-    breakpoint()
     if isinstance(target_height, SimulatedProbeResult):
         return target_height
     volumetric_capacity = get_well_volumetric_capacity(well_geometry)
