+++ api/src/opentrons/protocol_engine/state/wells.py
@@ -19,7 +19,10 @@ from opentrons.protocol_engine.types import (
     WellInfoSummary,
     WellLiquidInfo,
 )
-from opentrons.types import SimulatedProbeResult, LiquidTrackingType
+from opentrons.protocol_engine.types.liquid_level_detection import (
+    SimulatedProbeResult,
+    LiquidTrackingType,
+)
 
 from . import update_types
 from ._abstract_store import HasState, HandlesActions
@@ -83,7 +86,6 @@ class WellStore(HasState[WellState], HandlesActions):
             self._state.probed_volumes[labware_id] = {}
         updated_height = _none_from_clear(state_update.height)
         updated_volume = _none_from_clear(state_update.volume)
-        breakpoint()
         self._state.probed_heights[labware_id][well_name] = ProbedHeightInfo(
             height=updated_height,
             last_probed=state_update.last_probed,
